Add Search Icon in Bottom Naviagation Bar
