+++ flutter/ottugi_curry/lib/config/config.dart
@@ -6,7 +6,7 @@ import 'package:ottugi_curry/view/page/login/login_page.dart';
 import 'package:ottugi_curry/view/page/main/main_page.dart';
 import 'package:ottugi_curry/view/page/recipe_camera/recipe_camera_page.dart';
 import 'package:ottugi_curry/view/page/recipe_detail/recipe_detail_page_v2.dart';
-import 'package:ottugi_curry/view/page/search/search_page.dart';
+import 'package:ottugi_curry/view/page/text_search/text_search_page.dart';
 import 'package:ottugi_curry/view/page/user/user_page.dart';
 
 // token 저장하는 secure storage
@@ -50,9 +50,8 @@ class Config {
   ];
 
   // 레시피 상세 페이지에서 조리순서를 보는 방식: 음성/사진/즐글
-  static int soundView = 0;
-  static int galleryView = 1;
-  static int textListView = 2;
+  static int galleryView = 0;
+  static int textListView = 1;
 
   // storage 이름
   static String id = 'id';
