+++ flutter/ottugi_curry/lib/repository/lately_repository.dart
@@ -4,7 +4,7 @@ import 'package:dio/dio.dart';
 
 part 'lately_repository.g.dart';
 
-@RestApi(baseUrl: "http://192.168.151.183:8080")
+@RestApi(baseUrl: "http://192.168.0.51:8080")
 abstract class LatelyRepository {
   factory LatelyRepository(Dio dio, {String baseUrl}) = _LatelyRepository;
