+++ flutter/ottugi_curry/lib/repository/recipe_repository.dart
@@ -6,7 +6,7 @@ import 'package:dio/dio.dart';
 
 part 'recipe_repository.g.dart';
 
-@RestApi(baseUrl: "http://192.168.219.104:8080")
+@RestApi(baseUrl: "http://192.168.0.51:8080")
 abstract class RecipeRepository {
   factory RecipeRepository(Dio dio, {String baseUrl}) = _RecipeRepository;
