+++ flutter/ottugi_curry/lib/repository/user_repository.dart
@@ -4,7 +4,7 @@ import 'package:dio/dio.dart';
 
 part 'user_repository.g.dart';
 
-@RestApi(baseUrl: "http://192.168.219.104:8080")
+@RestApi(baseUrl: "http://192.168.0.51:8080")
 abstract class UserRepository {
   factory UserRepository(Dio dio, {String baseUrl}) = _UserRepository;
