+++ flutter/ottugi_curry/lib/view/page/bookmark/bookmark_list_page.dart
@@ -2,8 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:number_paginator/number_paginator.dart';
 import 'package:ottugi_curry/config/color_schemes.dart';
-import 'package:ottugi_curry/view/page/list/categories_widget.dart';
+import 'package:ottugi_curry/utils/long_string_to_list_utils.dart';
 import 'package:ottugi_curry/view/controller/bookmark/bookmark_controller.dart';
+import 'package:ottugi_curry/view/page/recipe_list/categories_widget.dart';
 
 class BookmrkListPage extends StatefulWidget {
   final String mode;
@@ -173,7 +174,7 @@ class BookmrkListPageState extends State<BookmrkListPage> {
                                 Row(children: [
                                   Expanded(
                                       child: Text(
-                                    '${menuItem.ingredients}',
+                                    '${extractOnlyContent(menuItem.ingredients ?? '')}',
                                     style:
                                         Theme.of(context).textTheme.bodySmall,
                                     overflow: TextOverflow.ellipsis,
